Remove duplicate mention detection code
